Move helper interface to msg.go
